requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
databases[mysql]
python-jose[cryptography]
passlib[bcrypt]
pandas
numpy
pygad
experta
pymysql
python-dotenv